+++ src/Loader/Loader.php
@@ -78,16 +78,6 @@ abstract class Loader implements LoaderInterface
         return true;
     }
 
-    /**
-     * Can be implemented in a child class to track how long a request waited for its response.
-     */
-    protected function trackRequestStart(?float $microtime = null): void {}
-
-    /**
-     * Can be implemented in a child class to track how long a request waited for its response.
-     */
-    protected function trackRequestEnd(?float $microtime = null): void {}
-
     protected function callHook(string $hook, mixed ...$arguments): void
     {
         if (!array_key_exists($hook, $this->hooks)) {